# docs/swagger.yaml
basePath: /
definitions:
  entities.Status:
    enum:
    - In Progress
    - Pending
    - Completed
    type: string
    x-enum-varnames:
    - InProgress
    - Pending
    - Completed
  entities.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/entities.Status'
      title:
        type: string
      updated_at:
        type: string
    type: object
  entities.UpdateTaskDTO:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - description
    - status
    - title
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A RESTful API for managing tasks
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Task Manager API
  version: "1.0"
paths:
  /tasks:
    get:
      description: Retrieve a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/entities.Task'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with title, description, and status
      parameters:
      - description: Task object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/entities.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created task
          schema:
            $ref: '#/definitions/entities.Task'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a task
      tags:
      - tasks
    get:
      description: Retrieve a specific task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task object
          schema:
            $ref: '#/definitions/entities.Task'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a task by ID
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Update an existing task by ID. All fields are optional.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update object
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateTaskDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated task
          schema:
            $ref: '#/definitions/entities.Task'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a task
      tags:
      - tasks
swagger: "2.0"